rag_local/requirements.txt
openai
beautifulsoup4
python-dotenv
langchain
langchain-openai
pypdf
pypdf2
langchain-hub
azure-storage-blob
azure-identity
pydrive.auth
pydrive.drive
pandas
langchain-community
google_auth_oauthlib.flow
googleapiclient.discovery 
pywin32
PyQt5
python-pptx
python-docx
markdown2
faiss-cpu
numpy
sentence-transformers
pymupdf
pytesseract
Pillow
fitz  # pymupdf
